[API] Rename campaignParticipation folder
In campaign-participation according to others
