+++ api/lib/application/campaign-participations/campaign-participation-controller.js
