+++ api/lib/application/campaign-participations/index.js
